% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dd.R
\name{all_revenue_dd}
\alias{all_revenue_dd}
\title{Data Dictionary for \code{all_revenue}}
\usage{
all_revenue_dd()
}
\value{
A \code{ptblank_informant} object.
}
\description{
The \code{all_revenue_dd()} function generates a data dictionary based on the
\code{all_revenue} table.
}
