% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dd.R
\name{all_sessions_dd}
\alias{all_sessions_dd}
\title{Data Dictionary for \code{all_sessions}}
\usage{
all_sessions_dd()
}
\value{
A \code{ptblank_informant} object.
}
\description{
The \code{all_sessions_dd()} function generates a data dictionary based on the
\code{all_sessions} table.
}
