% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_sj_all_revenue.R
\name{dd_sj_all_revenue}
\alias{dd_sj_all_revenue}
\title{Data Dictionary: sj_all_revenue}
\usage{
dd_sj_all_revenue()
}
\description{
The \code{dd_sj_all_revenue()} function generates a data dictionary based on the
\code{sj_all_revenue} table, which is available in the \strong{intendo} package as
a dataset.
}
