% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_sj_all_sessions.R
\name{dd_sj_all_sessions}
\alias{dd_sj_all_sessions}
\title{Data Dictionary: dd_sj_all_sessions}
\usage{
dd_sj_all_sessions()
}
\description{
The \code{dd_sj_all_sessions()} function generates a data dictionary based on the
\code{sj_all_sessions} table, which is available in the \strong{intendo} package as
a dataset.
}
