% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_sj_user_summary.R
\name{dd_sj_user_summary}
\alias{dd_sj_user_summary}
\title{Data Dictionary: sj_user_summary}
\usage{
dd_sj_user_summary()
}
\description{
The \code{dd_sj_user_summary()} function generates a data dictionary based on the
\code{sj_user_summary} table, which is available in the \strong{intendo} package as a
dataset.
}
