% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{intendo_daily_users}
\alias{intendo_daily_users}
\title{The \code{intendo_daily_users} dataset}
\format{
A tibble with 743306 rows and 14 variables:
\describe{
\item{user_id}{}
\item{session_id}{}
\item{time}{}
\item{total_sessions}{}
\item{total_time}{}
\item{level_reached}{}
\item{at_eoc}{}
\item{in_ftue}{}
\item{is_customer}{}
\item{iap_revenue}{}
\item{ad_revenue}{}
\item{total_revenue}{}
\item{acquired}{}
\item{iap_count}{}
\item{ad_count}{}
}
}
\usage{
intendo_daily_users
}
\description{
The \code{intendo_daily_users} dataset
}
\examples{
# Here is a glimpse at the data
# available in `intendo_daily_users`
dplyr::glimpse(intendo_daily_users)
}
\keyword{datasets}
