% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intendo_meta.R
\name{intendo_metadata}
\alias{intendo_metadata}
\title{Get a data frame with metadata for all datasets}
\usage{
intendo_metadata(include = c("all", "perfect", "faulty"))
}
\arguments{
\item{include}{Should both perfect and faulty datasets be included in the
metadata table? This is the default (\code{"all"}), otherwise one can use the
\code{"perfect"} or \code{"faulty"} keywords to get a subset of the metadata table.}
}
\value{
A data frame.
}
\description{
The \code{intendo_metadata()} function provides a data frame containing metadata
for datasets in the \strong{intendo} package.
}
\examples{

# Obtain metadata on all datasets in the package but
# only those in their 'perfect' form
intendo_metadata(include = "perfect")

}
