% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{intendo_revenue}
\alias{intendo_revenue}
\title{The \code{intendo_revenue} dataset}
\format{
A tibble with 1969139 rows and 8 variables:
\describe{
\item{user_id}{}
\item{session_id}{}
\item{time}{}
\item{name}{}
\item{size}{}
\item{type}{}
\item{price}{}
\item{revenue}{}
}
}
\usage{
intendo_revenue
}
\description{
The \code{intendo_revenue} dataset
}
\examples{
# Here is a glimpse at the data
# available in `intendo_revenue`
dplyr::glimpse(intendo_revenue)
}
\keyword{datasets}
