% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{intendo_users}
\alias{intendo_users}
\title{The \code{intendo_users} dataset}
\format{
A tibble with 743306 rows and 14 variables:
\describe{
\item{user_id}{A unique identifier for the user.}
\item{first_login}{A date-time that indicates the first time the player started the game.}
\item{iap_count}{}
\item{iap_revenue}{}
\item{ad_count}{}
\item{ad_revenue}{}
\item{total_revenue}{}
\item{customer}{}
\item{subscriber}{}
\item{first_iap}{}
\item{platform}{}
\item{device}{}
\item{acquired}{}
\item{country}{}
}
}
\usage{
intendo_users
}
\description{
The \code{intendo_users} dataset
}
\examples{
# Here is a glimpse at the data
# available in `intendo_users`
dplyr::glimpse(intendo_users)
}
\keyword{datasets}
