% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sj_all_revenue}
\alias{sj_all_revenue}
\title{All revenue amounts for Super Jetroid}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 214080 rows and 11 columns.
}
\usage{
sj_all_revenue
}
\description{
All revenue amounts for Super Jetroid
}
\examples{
# Here is a glimpse at the data
# available in `sj_all_revenue`
dplyr::glimpse(sj_all_revenue)


}
\keyword{datasets}
