% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sj_all_sessions}
\alias{sj_all_sessions}
\title{All player sessions for Super Jetroid}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 66751 rows and 9 columns.
}
\usage{
sj_all_sessions
}
\description{
All player sessions for Super Jetroid
}
\examples{
# Here is a glimpse at the data
# available in `sj_all_sessions`
dplyr::glimpse(sj_all_sessions)


}
\keyword{datasets}
