% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sj_all_sessions_f}
\alias{sj_all_sessions_f}
\title{All player sessions for Super Jetroid (faulty ver.)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 66756 rows and 9 columns.
}
\usage{
sj_all_sessions_f
}
\description{
All player sessions for Super Jetroid (faulty ver.)
}
\examples{
# Here is a glimpse at the data
# available in `sj_all_sessions_f`
dplyr::glimpse(sj_all_sessions_f)


}
\keyword{datasets}
