% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sj_user_summary}
\alias{sj_user_summary}
\title{User summaries for Super Jetroid}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1231 rows and 6 columns.
}
\usage{
sj_user_summary
}
\description{
User summaries for Super Jetroid
}
\examples{
# Here is a glimpse at the data
# available in `sj_user_summary`
dplyr::glimpse(sj_user_summary)


}
\keyword{datasets}
