% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sj_user_summary_f}
\alias{sj_user_summary_f}
\title{User summaries for Super Jetroid (faulty ver.)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1231 rows and 6 columns.
}
\usage{
sj_user_summary_f
}
\description{
User summaries for Super Jetroid (faulty ver.)
}
\examples{
# Here is a glimpse at the data
# available in `sj_user_summary_f`
dplyr::glimpse(sj_user_summary_f)


}
\keyword{datasets}
