% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sj_users_daily}
\alias{sj_users_daily}
\title{Daily users playing Super Jetroid}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 48341 rows and 23 columns.
}
\usage{
sj_users_daily
}
\description{
Daily users playing Super Jetroid
}
\examples{
# Here is a glimpse at the data
# available in `sj_users_daily`
dplyr::glimpse(sj_users_daily)


}
\keyword{datasets}
