% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sj_users_daily_f}
\alias{sj_users_daily_f}
\title{Daily users playing Super Jetroid (faulty ver.)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 48485 rows and 23 columns.
}
\usage{
sj_users_daily_f
}
\description{
Daily users playing Super Jetroid (faulty ver.)
}
\examples{
# Here is a glimpse at the data
# available in `sj_users_daily_f`
dplyr::glimpse(sj_users_daily_f)


}
\keyword{datasets}
