% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dd.R
\name{user_summary_dd}
\alias{user_summary_dd}
\title{Data Dictionary for \code{user_summary}}
\usage{
user_summary_dd()
}
\description{
The \code{user_summary_dd()} function generates a data dictionary based on the
\code{user_summary} table.
}
