% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dd.R
\name{users_daily_dd}
\alias{users_daily_dd}
\title{Data Dictionary for \code{users_daily}}
\usage{
users_daily_dd()
}
\value{
A \code{ptblank_informant} object.
}
\description{
The \code{users_daily_dd()} function generates a data dictionary based on the
\code{users_daily} table.
}
