% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_sj_users_daily.R
\name{dd_sj_users_daily}
\alias{dd_sj_users_daily}
\title{Data Dictionary: sj_users_daily}
\usage{
dd_sj_users_daily()
}
\description{
The \code{dd_sj_users_daily()} function generates a data dictionary based on the
\code{sj_users_daily} table, which is available in the \strong{intendo} package as a
dataset.
}
